% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{index}
\alias{index}
\title{Pharmacological interventions for chronic obstructive pulmonary disease}
\format{
A data-frame with 21 rows of arm-based data and 17 columns (maximum
   number of 4 arms).
}
\usage{
data(index)
}
\description{
A dataset of 21 trials comparing seven pharmacologic
  interventions with each other and placebo in patients with chronic
  obstructive pulmonary disease (COPD). The exacerbation of COPD is the
  analysed binary outcome.
}
\details{
The interventions have been coded as follows: 1, placebo; 2,
  budesonide; 3, budesonide plus formoterol; 4, fluticasone; 5, fluticasone
  plus salmeterol; 6, formoterol; 7, salmeterol; and 8, tiotropium.
}
\keyword{datasets}
