% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dataset.index_function.R
\name{get.dataset.index}
\alias{get.dataset.index}
\title{Get the index of specific dataset(s)}
\usage{
get.dataset.index(pmid, year = NULL)
}
\arguments{
\item{pmid}{A scalar or vector with the PMID number of the systematic
reviews found in the database.}

\item{year}{A scalar to define the year publication of interest.
\code{get.dataset.index} will return all systematic reviews found in the
database under this year.}
}
\value{
\code{get.dataset.index} returns a vector (when \code{pmid} is scalar) or a
  data-frame (when \code{pmid} is a vector) with the following
  characteristics (column names):
  \tabular{ll}{
   \strong{nmadb.ID} \tab The ID number as provided in the R package
   \href{https://CRAN.R-project.org/package=nmadb}{nmadb}.\cr
   \tab \cr
   \strong{PMID} \tab The PMID number.\cr
   \tab \cr
   \strong{First.Author} \tab The name of the first author.\cr
   \tab \cr
   \strong{Year} \tab The year of publication.\cr
   \tab \cr
   \strong{Journal.Name} \tab The abbreviated name of the journal.\cr
   \tab \cr
   \strong{Title} \tab The title of the systematic review.\cr
   \tab \cr
   \strong{Outcome.Type} \tab The outcome type as suggested by Turner et al.
   (2012) and distinguished into objective, semi-objective and subjective.\cr
   \tab \cr
   \strong{Intervention.Comparison.Type} \tab The treatment-comparator type
   as suggested by Turner et al. (2012) and distinguished into
   pharmacological versus placebo, pharmacological versus pharmacological,
   and non-pharmacological versus any.\cr
   \tab \cr
   \strong{Includes.ToC.where} \tab Whether the extracted study-level
   aggregate characteristics were found in the main article, Appendix or
   both.\cr
   \tab \cr
   \strong{Source.ToC} \tab The exact location in the systematic review
   where the extracted study-level aggregate characteristics were found, such
   as Table X in the main article, and/ or Appendix.\cr
   \tab \cr
   }
}
\description{
A function to retrieve the characteristics of one of more systematic
  reviews found in the database using their PMID number or year of
  publication.
}
\details{
When \code{year} is specified, the argument \code{pmid} is ignored, and
  \code{get.dataset.index} returns all systematic reviews found in the
  database under this year.

  Except for \strong{PMID}, \strong{Includes.ToC.where}, and
  \strong{Source.ToC}, all other characteristics were retrieved from the R
  package \href{https://CRAN.R-project.org/package=nmadb}{nmadb} using the
  function \code{\link[nmadb:getNMADB]{getNMADB}} and subsetting to those
  systematic reviews with available data
  (i.e., \code{Data.available == "True"} when using the function
  \code{\link[nmadb:getNMADB]{getNMADB}}). The database was reduced further
  during extraction, for instance, due to the reporting quality and data of
  the systematic review relating to the extraction process.
}
\examples{
# For one systematic review with PMID number 25626481.
get.dataset.index(pmid = 25626481)

# For two systematic reviews with PMID numbers 25626481 and 21328309,
# respectively.
get.dataset.index(pmid = c(25626481, 21328309))

# For systematic reviews published during 2010. The argument 'pmid' does not
# need to be specified.
get.dataset.index(year = 2010)

}
\references{
Turner RM, Davey J, Clarke MJ, Thompson SG, Higgins JP. Predicting the extent
of heterogeneity in meta-analysis, using empirical data from the Cochrane
Database of Systematic Reviews.
\emph{Int J Epidemiol} 2012;\bold{41}(3):818--27. doi: 10.1093/ije/dys041.
}
\seealso{
\code{\link[nmadb:getNMADB]{getNMADB}}
}
\author{
{Loukia M. Spineli}
}
