% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dataset_function.R
\name{get.dataset}
\alias{get.dataset}
\title{Get the selected dataset}
\usage{
get.dataset(pmid)
}
\arguments{
\item{pmid}{A scalar with the PMID number of the systematic review found in
the database.}
}
\value{
\code{get.dataset} returns a data-frame with rows referring to the studies
  and columns to the aggregate study-level characteristics as extracted from
  the report of the corresponding systematic review.
}
\description{
Function to retrieve the dataset with aggregate study-level clinical and
  methodological characteristics (that may act as effect modifiers) extracted
  from the selected systematic review of the database.
}
\details{
Each dataset refers to a connected network for a specific outcome. The R
  package \href{https://CRAN.R-project.org/package=nmadb}{nmadb} was used to
  retrieve the corresponding dataset. Specifically, the function
  \code{\link[nmadb:readByID]{readByID}} was employed to download the dataset
  in the long format. Then, the function
  \code{\link[netmeta:pairwise]{pairwise}} of the R package
  \href{https://CRAN.R-project.org/package=netmeta}{netmeta} was implemented
  to convert the dataset into wide format with each row repeated as many
  times as the number of possible comparisons made in the corresponding
  study. The study names (or references) and treatment comparisons, as
  returned by \code{\link[nmadb:readByID]{readByID}}, were used to locate the
  studies in the corresponding report of the systematic review (and Appendix,
  if available) and extract the characteristics available in the relevant
  table(s). Each characteristic occupies one column in the dataset.
  Characteristics pertaining to intervention features occupied one column for
  the experimental and another for the control treatment in the corresponding
  comparison.
}
\examples{
# Get dataset for systematic review with PMID number 25626481
get.dataset(pmid = 25626481)

}
\seealso{
\code{\link[nmadb:getNMADB]{getNMADB}},
  \code{\link[netmeta:pairwise]{pairwise}}
}
\author{
{Loukia M. Spineli}
}
